% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_list_projects}
\alias{auth_list_projects}
\title{List available DNAnexus projects}
\usage{
auth_list_projects()
}
\value{
A character vector of project names and IDs.
}
\description{
Returns a list of all projects accessible to the current user.
}
\examples{
\dontrun{
auth_list_projects()
}
}
