% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_logout}
\alias{auth_logout}
\title{Logout from DNAnexus}
\usage{
auth_logout()
}
\value{
Invisible TRUE on success.
}
\description{
Clears the local DNAnexus authentication token and session.
}
\examples{
\dontrun{
auth_logout()
}
}
