% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_status}
\alias{auth_status}
\title{Check current DNAnexus authentication status}
\usage{
auth_status()
}
\value{
A named list with \code{user} and \code{project}.
}
\description{
Returns the current logged-in user and selected project.
}
\examples{
\dontrun{
auth_status()
}
}
