% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ukbflow-package}
\alias{ukbflow}
\alias{ukbflow-package}
\title{ukbflow: Streamlined Workflow for UK Biobank Data Extraction, Analysis, and Visualization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an R-native, RAP-aware workflow system for UK Biobank controlled-data analysis on the Research Analysis Platform (RAP). Includes tools for phenotype extraction and decoding, variable derivation, survival and association analysis, genetic risk score computation, audit records, and publication-quality visualization. For details on the UK Biobank resource, see Bycroft et al. (2018) \doi{10.1038/s41586-018-0579-z}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/evanbio/ukbflow}
  \item \url{https://evanbio.github.io/ukbflow/}
  \item Report bugs at \url{https://github.com/evanbio/ukbflow/issues}
}

}
\author{
\strong{Maintainer}: Yibin Zhou \email{evanzhou.bio@gmail.com} (\href{https://orcid.org/0009-0009-4600-8175}{ORCID})

}
\keyword{internal}
