% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_logout}
\alias{auth_logout}
\title{Logout from DNAnexus}
\usage{
auth_logout()
}
\value{
Invisible TRUE on success.
}
\description{
Invalidates the current DNAnexus session on the remote platform. The local
token file is not removed but becomes invalid. A new token must be generated
from the DNAnexus platform before calling \code{\link[=auth_login]{auth_login()}} again.
}
\examples{
\dontrun{
auth_logout()
}
}
