% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_select_project}
\alias{auth_select_project}
\title{Select a DNAnexus project}
\usage{
auth_select_project(project)
}
\arguments{
\item{project}{(character) Project ID in the form \code{"project-XXXXXXXXXXXX"}.}
}
\value{
Invisible TRUE on success.
}
\description{
Switches the active project context on the DNAnexus platform. Only project
IDs (e.g. \code{"project-XXXXXXXXXXXX"}) are accepted. Run
\code{\link[=auth_list_projects]{auth_list_projects()}} to find your project ID.
}
\examples{
\dontrun{
auth_select_project("project-XXXXXXXXXXXX")
}
}
