% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ukbflow-package}
\alias{ukbflow}
\alias{ukbflow-package}
\title{ukbflow: Streamlined Workflow for UK Biobank Data Extraction, Analysis, and Visualization}
\description{
An R package for UK Biobank cloud-based analysis on the Research Analysis Platform (RAP). Provides tools for phenotype extraction and decoding, variable derivation, survival and association analysis, genetic risk score computation, and publication-quality visualization.
}
\author{
\strong{Maintainer}: Evan Zhou \email{evanzhou.bio@gmail.com}

}
\keyword{internal}
