% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ukbflow-package}
\alias{ukbflow}
\alias{ukbflow-package}
\title{ukbflow: Streamlined Workflow for UK Biobank Data Extraction, Analysis, and Visualization}
\description{
A streamlined workflow for UK Biobank data extraction, clinical analysis, and publication-quality visualization. Covers upstream data cleaning and field ID mapping, midstream clinical and association analysis, and downstream high-quality plot generation.
}
\author{
\strong{Maintainer}: Evan Zhou \email{evanzhou.bio@gmail.com}

}
\keyword{internal}
